# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Veterinary Management
  description: Veterinary Management
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
    description: "Local server"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication using OAuth2 Bearer token.

        **How to authenticate:**
        1. POST to `/loginJWT` with `administrator_id` and `password`
        2. Receive JWT token in response
        3. Include token in Authorization header: `Bearer <token>`

        **Available scopes:**
        - `RECEPTIONIST` - Clinic receptionist staff (manage visits, view schedules)
        - `CLINIC_MANAGER` - Clinic manager (full system access)
        - `VETERINARIAN` - Veterinarian (manage own schedule, view assigned visits)
        - `PET_OWNER` - Pet owner (view own pets and visits)
  parameters:
...
    promotion_id:
      name: promotion_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/PromotionId'
    discount_id:
      name: discount_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/DiscountId'
    loyalty_tier_id:
      name: loyalty_tier_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/LoyaltyTierId'
...
  requestBodies:
...
    PromotionInput:
      description: "Promotion object input"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
              discount_code:
                type: string
              start_date:
                $ref: '#/components/schemas/LocalDate'
              end_date:
                $ref: '#/components/schemas/LocalDate'
    DiscountInput:
      description: "Discount object input"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
              discount_type:
                $ref: '#/components/schemas/DiscountType'
              value_amount:
                type: number
              start_date:
                $ref: '#/components/schemas/LocalDate'
              end_date:
                $ref: '#/components/schemas/LocalDate'
              max_uses:
                type: integer
              uses_count:
                type: integer
...
    LoyaltyTierInput:
      description: "Loyalty tier input"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
              min_points:
                type: integer
              benefits:
                type: string
...
  schemas:
...
    LocalDate:
      type: string
      format: date
      description: "Local date in the format YYYY-MM-DD"
      default: "2025-01-01"
...
    PromotionId:
      type: integer
      format: int64
      description: "Unique identifier for a Promotion"
    DiscountId:
      type: integer
      format: int64
      description: "Unique identifier for a Discount"
    DiscountType:
      type: string
      enum:
        - PERCENTAGE
        - FIXED_AMOUNT
        - LOYALTY_TIER
      description: "Type of discount"
...
    LoyaltyTierId:
      type: integer
      format: int64
      description: "Unique identifier for a Loyalty Tier"
...
    Promotion:
      type: object
      properties:
        promotion_id:
          $ref: '#/components/schemas/PromotionId'
        name:
          type: string
        description:
          type: string
        discount_code:
          type: string
        start_date:
          $ref: '#/components/schemas/LocalDate'
        end_date:
          $ref: '#/components/schemas/LocalDate'
    Discount:
      type: object
      properties:
        discount_id:
          $ref: '#/components/schemas/DiscountId'
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        value_amount:
          type: number
        start_date:
          $ref: '#/components/schemas/LocalDate'
        end_date:
          $ref: '#/components/schemas/LocalDate'
        max_uses:
          type: integer
        uses_count:
          type: integer
...
    LoyaltyTier:
      type: object
      properties:
        loyalty_tier_id:
          $ref: '#/components/schemas/LoyaltyTierId'
        name:
          type: string
        description:
          type: string
        min_points:
          type: integer
        benefits:
          type: string
        discount_id:
          $ref: '#/components/schemas/DiscountId'
...
paths:
...
  /promotions:
    post:
      tags: [ Promotions ]
      summary: "Create a new Promotion"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      requestBody:
        $ref: '#/components/requestBodies/PromotionInput'
      responses:
        '201':
          description: "Promotion created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promotion'
        '400':
          description: "Invalid input"
  /promotions/{promotion_id}:
    get:
      tags: [ Promotions ]
      summary: "Get a promotion"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      parameters:
        - $ref: '#/components/parameters/promotion_id'
      responses:
        '200':
          description: "Promotion found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promotion'
        '404':
          description: "Promotion not found"
  /promotions/{promotion_id}/discounts:
    post:
      tags: [ Promotions ]
      summary: "Create a new discount for a Promotion"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      parameters:
        - $ref: '#/components/parameters/promotion_id'
      requestBody:
        $ref: '#/components/requestBodies/DiscountInput'
      responses:
        '201':
          description: "Discount created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Discount'
        '400':
          description: "Invalid input"
        '404':
          description: "Promotion not found"
    get:
      tags: [ Promotions ]
      summary: "Get all discounts from a Promotion"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      parameters:
        - $ref: '#/components/parameters/promotion_id'
      responses:
        '200':
          description: "Discounts found"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Discount'
        '404':
          description: "Promotion not found"
    delete:
      tags: [ Promotions ]
      summary: "Delete all discounts from a Promotion"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      parameters:
        - $ref: '#/components/parameters/promotion_id'
      responses:
        '204':
          description: "Discounts deleted"
        '404':
          description: "Promotion not found"
  /promotions/{promotion_id}/discounts/{discount_id}:
    get:
      tags: [ Promotions ]
      summary: "Get a discount"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      parameters:
        - $ref: '#/components/parameters/discount_id'
      responses:
       '200':
         description: "Discount found"
         content:
           application/json:
             schema:
              $ref: '#/components/schemas/Discount'
       '404':
          description: "Discount not found"
    put:
      tags: [ Promotions ]
      summary: "Update a discount"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      parameters:
        - $ref: '#/components/parameters/discount_id'
      requestBody:
        $ref: '#/components/requestBodies/DiscountInput'
      responses:
        '200':
          description: "Discount updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Discount'
        '400':
          description: "Invalid input"
        '404':
          description: "Discount not found"
  /promotions/{promotion_id}/discounts/{discount_id}/loyaltyTiers:
    post:
      tags: [ Loyalty Tiers ]
      summary: "Create a new Loyalty Tier"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      parameters:
        - $ref: '#/components/parameters/discount_id'
      requestBody:
        $ref: '#/components/requestBodies/LoyaltyTierInput'
      responses:
        '201':
          description: "Loyalty Tier created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyTier'
        '400':
          description: "Invalid input"
        '404':
          description: "Discount not found"
    get:
      tags: [ Loyalty Tiers ]
      summary: "Get all Loyalty Tiers from a Discount"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      parameters:
        - $ref: '#/components/parameters/discount_id'
      responses:
        '200':
          description: "Loyalty Tiers found"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoyaltyTier'
        '404':
          description: "Discount not found"
  /promotions/{promotion_id}/discounts/{discount_id}/loyaltyTiers/{loyalty_tier_id}:
    get:
      tags: [ Loyalty Tiers ]
      summary: "Get a Loyalty Tier"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      parameters:
        - $ref: '#/components/parameters/loyalty_tier_id'
      responses:
        '200':
          description: "Loyalty Tier found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyTier'
        '404':
          description: "Loyalty Tier not found"
    put:
      tags: [ Loyalty Tiers ]
      summary: "Update a Loyalty Tier"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      parameters:
        - $ref: '#/components/parameters/loyalty_tier_id'
      requestBody:
        $ref: '#/components/requestBodies/LoyaltyTierInput'
      responses:
        '200':
          description: "Loyalty Tier updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyTier'
        '400':
          description: "Invalid input"
        '404':
          description: "Loyalty Tier not found"
    delete:
      tags: [ Loyalty Tiers ]
      summary: "Delete a Loyalty Tier"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      parameters:
        - $ref: '#/components/parameters/loyalty_tier_id'
      responses:
        '204':
          description: "Loyalty Tier deleted"
        '404':
          description: "Loyalty Tier not found"

  # ============================================
  # Visits
  # ============================================
...